Made the component parsers more generic to decouple them from METARs, allowing them to be used by TAF among others.
