+++ Source/AviationWeather/Source/components.cpp
@@ -1,17 +1,17 @@
 /**********************************************************************************
-*                                                                                *
-* Copyright (c) 2015 Steven Frost, Orion Lyau. All rights reserved.              *
-*                                                                                *
-* This source is subject to the MIT License.                                     *
-* See http://opensource.org/licenses/MIT                                         *
-*                                                                                *
-* THIS CODE AND INFORMATION IS PROVIDED "AS IS" WITHOUT WARRANTY OF ANY KIND,    *
-* EITHER EXPRESSED OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE IMPLIED          *
-* WARRANTIES OF MERCHANTABILITY AND/OR FITNESS FOR A PARTICULAR PURPOSE.         *
-*                                                                                *
-* NOT TO BE USED AS A SOLE SOURCE OF INFORMATION FOR FLIGHT CRITICAL OPERATIONS. *
-*                                                                                *
-**********************************************************************************/
+ *                                                                                *
+ * Copyright (c) 2015 Steven Frost, Orion Lyau. All rights reserved.              *
+ *                                                                                *
+ * This source is subject to the MIT License.                                     *
+ * See http://opensource.org/licenses/MIT                                         *
+ *                                                                                *
+ * THIS CODE AND INFORMATION IS PROVIDED "AS IS" WITHOUT WARRANTY OF ANY KIND,    *
+ * EITHER EXPRESSED OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE IMPLIED          *
+ * WARRANTIES OF MERCHANTABILITY AND/OR FITNESS FOR A PARTICULAR PURPOSE.         *
+ *                                                                                *
+ * NOT TO BE USED AS A SOLE SOURCE OF INFORMATION FOR FLIGHT CRITICAL OPERATIONS. *
+ *                                                                                *
+ **********************************************************************************/
 
 #include "AviationWeatherPch.h"
